pyproject.toml
[project]
name = "Ifc"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "ThiagoMonteiroBSI", email = "thiagodepaulamonteiro@gmail.com"},
]
dependencies = ["django>=6.0.2", "djangorestframework>=3.16.1"]
requires-python = ">=3.14"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
